+++ Modules/sessionFrame.lua
@@ -4,7 +4,7 @@
 -- Create Date : 1/20/2015 3:48:38 AM
 
 local addon = LibStub("AceAddon-3.0"):GetAddon("RCLootCouncil")
-local RCSessionFrame = addon:NewModule("RCSessionFrame", "AceTimer-3.0")
+local RCSessionFrame = addon:NewModule("RCSessionFrame", "AceTimer-3.0", "AceEvent-3.0")
 local ST = LibStub("ScrollingTable")
 local L = LibStub("AceLocale-3.0"):GetLocale("RCLootCouncil")
 
@@ -33,12 +33,14 @@ end
 
 function RCSessionFrame:OnEnable()
 	addon:Debug("RCSessionFrame", "enabled")
+	self:RegisterMessage("RCLootStatusReceived", "UpdateLootStatus")
 	ml = addon:GetActiveModule("masterlooter")
 end
 
 function RCSessionFrame:OnDisable()
 	self.frame:Hide()
 	self.frame.rows = {}
+	self:UnregisterMessage("RCLootStatusReceived")
 	awardLater = false
 end
 
@@ -113,6 +115,20 @@ function RCSessionFrame:Update()
 	else
 		self.frame.startBtn:SetText(_G.START)
 	end
+end
+
+function RCSessionFrame:UpdateLootStatus ()
+	if not self.frame then return end
+	local status, list = addon:GetLootStatusData()
+	self.frame.lootStatus:SetText(status)
+	self.frame.lootStatus:SetScript("OnEnter", function()
+		GameTooltip:SetOwner(self.frame.lootStatus, "ANCHOR_RIGHT")
+		GameTooltip:AddLine(L["Loot Status"])
+		for _, v in ipairs(list) do
+			GameTooltip:AddDoubleLine(addon:GetUnitClassColoredName(v.name), v.text)
+		end
+		GameTooltip:Show()
+	end)
 end
 
 function RCSessionFrame:DeleteItem(session, row)
@@ -217,6 +233,16 @@ function RCSessionFrame:GetFrame()
 		self:Disable()
 	end)
 	f.closeBtn = b2
+
+	-- Loot Status
+	f.lootStatus = addon.UI:New("Text", f.content, " ")
+	f.lootStatus:SetTextColor(1,1,1,1) -- White for now
+	f.lootStatus:SetHeight(20)
+	f.lootStatus:SetWidth(75)
+	-- f.lootStatus:SetPoint("BOTTOMRIGHT", f, "BOTTOMRIGHT", -10, 14)
+	f.lootStatus:SetPoint("LEFT", f.closeBtn, "RIGHT", 13, 1)
+	f.lootStatus:SetScript("OnLeave", addon.Utils.HideTooltip)
+	f.lootStatus.text:SetJustifyH("LEFT")
 
 	local st = ST:CreateST(self.scrollCols, 5, ROW_HEIGHT, nil, f.content)
 	st.frame:SetPoint("TOPLEFT",f,"TOPLEFT",10,-20)