+++ changelog.md
@@ -1,3 +1,16 @@
+### v2.10.1
+---
+**Loot Status**
+* Added loot status to the session frame.
+* This allows the ML to see how many have looted the boss before starting a session.
+
+* Added `notes` as an option to announce awards.
+
+
+###### Bugfixes
+* *Fixed random issue with realm name (#385).*
+
+
 ### v2.10.0
 ---
 * **Version Checker**